underpants.adoc - make asciidoctor happy with nested numbered list
It was complaining with an error:

  asciidoctor: WARNING: underpants.adoc: line 28: list item index:
  expected 2, got 3

So use something it's happier with. However, its numbering styles
don't exactly do what I want - and it doesn't seem easy to coerce
that, so use the default (alphabetic sub points) and put the top-level
point two back (so that the sub points have a parent and the numbering
increments correctly).